managers-service/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
pytest
requests>=2.0
PyJWT>=2.0
psycopg2-binary>=2.9
